+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `pybann/network.py` containing network methods
 - `tests/test_network.py`
 - `pybann/backpropagation.py` containing the backpropagation algorithm
+- `tests/backpropagation.py`
+- `data/iris/iris.data` & `data/iris/iris.names`
+- `examples/iris-example.py`
 
 ### Changed